fee tier validation

+++ programs/clearing_house/src/math/fees.rs
@@ -161,9 +161,9 @@ fn calculate_filler_reward(
     // for sufficiently small-sized order, reward based on fraction of fee paid
 
     let size_filler_reward = fee
-        .checked_mul(filler_reward_structure.reward_numerator)
+        .checked_mul(filler_reward_structure.reward_numerator as u128)
         .ok_or_else(math_error!())?
-        .checked_div(filler_reward_structure.reward_denominator)
+        .checked_div(filler_reward_structure.reward_denominator as u128)
         .ok_or_else(math_error!())?;
 
     let min_time_filler_reward = filler_reward_structure.time_based_reward_lower_bound;
